Introduce a new stored property on `RenderBlockContent.Aside` called `name` to stored the name or title of aside blocks. This allows DocC and applications that link to SwiftDocC to independently save and control the styling and the name of asides. Previously DocC conflated the style of the aside block (e.g. colors or other styling) with the name of the aside.
Aside now has four available initializers:
* init(style: AsideStyle, content: [RenderBlockContent])
* init(name: String, content: [RenderBlockContent])
* init(style: AsideStyle, name: String, content: [RenderBlockContent])
* init(asideKind: Markdown.Aside.Kind, content: [RenderBlockContent])

Most often DocC will use the fourth initializer to create aside blocks
from markdown, via RenderContentCompiler. This initializer, creating an
aside from a Swift Markdown Aside.Kind, also contains special logic to
determine a capitalized name for the aside that was previously
implemented in AsideStyle.

Simplify the `RenderBlockContent.AsideStyle` structure to act as if it
were an enum of the known styles supported by DocC Render: note, tip,
experiment, important, or warning. (Since this is public API, this
change retains AsideStyle as a structure with a string raw value.)

Also simplify the JSON encoding and decoding of Aside and AsideStyle
structures to save and read the style and name in a natural way.